+++ tests/test_reader_deprecations.py
@@ -15,31 +15,6 @@ def test_entry_error_url(cls):
         assert exc.url == 'feed'
 
 
-@pytest.mark.parametrize('name', ['update_feeds', 'update_feeds_iter'])
-def test_update_feeds_new_only(reader, name):
-    reader._parser = Parser()
-    method = getattr(reader, name)
-
-    with pytest.raises(TypeError):
-        list(method(new=True, new_only=True) or ())
-
-    class MyError(Exception):
-        pass
-
-    def get_feeds_for_update(url=None, new=None, enabled_only=False):
-        raise MyError(new)
-
-    reader._storage.get_feeds_for_update = get_feeds_for_update
-
-    with pytest.raises(MyError) as excinfo, pytest.deprecated_call():
-        list(method(new_only=True) or ())
-    assert excinfo.value.args[0] is True
-
-    with pytest.raises(MyError) as excinfo, pytest.deprecated_call():
-        list(method(new_only=False) or ())
-    assert excinfo.value.args[0] is None
-
-
 def test_updated_feed_updated():
     feed = UpdatedFeed('url', 1, 2)